Instructor edits: fixed printing connection parameters, fixed state machine.

+++ src/ble.c
@@ -117,10 +117,11 @@ void handle_ble_event(sl_bt_msg_t *evt){
 
     /**Call when parameters are set**/
     case sl_bt_evt_connection_parameters_id:
-#if (PRINT_PARAMS)
-      LOG_INFO("interval=%d, latency=%d, timeout=%d",(evt->data.evt_connection_parameters.interval*1.25),
-               evt->data.evt_connection_parameters.latency,
-               evt->data.evt_connection_parameters.timeout );
+#if (PRINT_PARAMS) // DOS: compiler error in formatting these parameters, included with ( ) and typecast to int
+      LOG_INFO("interval=%d, latency=%d, timeout=%d",
+               (int) (evt->data.evt_connection_parameters.interval*1.25),
+               (int) (evt->data.evt_connection_parameters.latency),
+               (int) (evt->data.evt_connection_parameters.timeout) );
 #endif
       break;
